fix wsl (ubuntu) wsl-open

+++ debian.sh
@@ -290,7 +290,7 @@ GNULINUX() {
 		WSL="[WSL(win10的linux子系统)]"
 		WINDOWSDISTRO='WSL'
 		#/usr/local/bin/wsl-open
-		if [ ! -e /usr/local/bin/wsl-open ]; then
+		if [ ! -e /usr/local/bin/wsl-open ] && [ ! -e usr/bin/wsl-open ]; then
 			echo '检测到您使用的是WSL,正在为您安装nodejs、npm和npm模块（wsl-open）'
 			apt install -y nodejs
 			wget -O- https://npmjs.org/install.sh | bash
